Multi-Server Phase 3 (Mostly Completed)

+++ mrbBase/Files.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Windows.Forms;
@@ -10,115 +12,37 @@ namespace mrbBase
 {
     public static class Files
     {
-        public const string MxdbFileDB = "I12.mhd";
-        public const string MxdbFileLevels = "Levels.mhd";
+        public const string MxdbFileDb = "I12.mhd";
         public const string MxdbFileNLevels = "NLevels.mhd";
         public const string MxdbFileRLevels = "RLevels.mhd";
         public const string MxdbFileMaths = "Maths.mhd";
         public const string MxdbFileEClasses = "EClasses.mhd";
         public const string MxdbFileOrigins = "Origins.mhd";
-        public const string MxdbFileSetTypes = "SetTypes.mhd";
         public const string MxdbFileSalvage = "Salvage.mhd";
         public const string MxdbFileRecipe = "Recipe.mhd";
-        public const string MxdbFileEnhDB = "EnhDB.mhd";
+        public const string MxdbFileEnhDb = "EnhDB.mhd";
         public const string MxdbFileBbCodeUpdate = "BBCode.mhd";
         public const string MxdbFileOverrides = "Compare.mhd";
         public const string MxdbFileModifiers = "AttribMod.mhd";
-        public const string MxdbFileLookups = "AttribLookup.mhd";
         public const string MxdbFileEffectIds = "GlobalMods.mhd";
         public const string MxdbFileGraphics = "I9.mhd";
+
         public const string JsonFileModifiers = "AttribMod.json";
-        //public const string PatchRtf = "patch.rtf";
-        private const string MxdbFileConfig = "Config.mhd";
+        public const string JsonFileTypeGrades = "TypeGrades.json";
         private const string JsonFileConfig = "Config.json";
-        private const string JsonFileConfigSP = "ConfigSP.json";
+
+        //public const string PatchRtf = "patch.rtf";
         private const string MxdbPowersReplTable = "PowersReplTable.mhd";
 
         public const string RoamingFolder = "Data\\";
+        public const string BuildsFolder = "Hero & Villain Builds\\";
+
         public static string FileData = string.Empty;
-
-        public static string FNameJsonConfig => Path.Combine(Path.Combine(GetAssemblyLoc(), RoamingFolder), JsonFileConfig);
-
-        public static string GetConfigSpFile()
-        {
-            var sPath = $@"{Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData)}\RebornTeam\";
-            return Path.Combine(sPath, JsonFileConfigSP);
-        }
-
-        public static string FDefaultPath => Path.Combine(Path.Combine(GetAssemblyLoc(), RoamingFolder), "Homecoming\\");
-
-        private static string FNameConfig => Path.Combine(Path.Combine(GetAssemblyLoc(), RoamingFolder), MxdbFileConfig);
-
+        public static string FNameJsonConfig => Path.Combine(AppContext.BaseDirectory, RoamingFolder, JsonFileConfig);
+        public static string FDefaultPath => Path.Combine(AppContext.BaseDirectory, RoamingFolder, "Generic\\");
         public static string FNamePowersRepl => Path.Combine(FPathAppData, MxdbPowersReplTable);
-
         private static string FPathAppData => MidsContext.Config != null ? MidsContext.Config.DataPath : FDefaultPath;
-
-        internal static string SearchUp(string folder, string fn)
-        {
-            string SearchUpRec(string foldername, string filename)
-            {
-                // if it is not properly rooted, give up, and use the path
-                if (!Path.IsPathRooted(filename))
-                {
-                    Console.WriteLine(filename);
-                    return filename;
-                }
-
-                // get the directory that holds the filename, filename should be a FULL path
-                var fnDir = Path.GetDirectoryName(filename);
-                // if the filename is already in a folder with the correct foldername, we need to go up twice instead of once.
-                string targetRoot;
-                if (fnDir.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar).EndsWith(foldername))
-                {
-                    targetRoot = Path.GetDirectoryName(fnDir);
-                    Console.WriteLine($"Targetroot: {targetRoot}");
-                }
-                else
-                {
-                    targetRoot = fnDir;
-                    Console.WriteLine($"Targetroot: {targetRoot}");
-                }
-
-                var parent = Path.GetDirectoryName(targetRoot);
-                Console.WriteLine($"Parent: {parent}");
-                if (parent == null) return null;
-                var attempt = Path.Combine(Path.Combine(parent, foldername), Path.GetFileName(filename));
-                Console.WriteLine($"attempt: {attempt}");
-                if (File.Exists(attempt))
-                    return attempt;
-                //if (Path.IsPathRooted(attempt)) return filename;
-                var recursed = SearchUpRec(foldername, attempt);
-                Console.WriteLine($"recursed: {recursed}");
-                if (recursed != null && File.Exists(recursed)) return recursed;
-                return File.Exists(filename) ? filename : null;
-            }
-
-            try
-            {
-                var result = SearchUpRec(folder, fn);
-                return result != null && File.Exists(result) ? result : fn;
-            }
-            catch
-            {
-                return fn;
-            }
-        }
-
-        /*public static string SelectDataFileLoad(string iDataFile)
-        {
-            var str = Path.Combine(FPathAppData, iDataFile);
-            if (Debugger.IsAttached)
-                str = SearchUp("Data", str);
-            FileData = FileData + str + '\n';
-            return str;
-        }*/
-
-        public static string GetAssemblyLoc()
-        {
-            var asmLoc = Assembly.GetExecutingAssembly().Location;
-            var dirLoc = $"{Directory.GetParent(asmLoc)}\\";
-            return dirLoc;
-        }
+        public static string FDefaultBuildsPath => Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), BuildsFolder);
 
         public static string SelectDataFileLoad(string iDataFile, string iPath = "")
         {
@@ -139,75 +63,27 @@ namespace mrbBase
             {
                 filePath = Path.GetFullPath(Path.Combine(Path.Combine(AppContext.BaseDirectory, @"..\..\"), RoamingFolder, DatabaseAPI.DatabaseName, iDataFile));
                 if (!Directory.Exists(FileIO.StripFileName(filePath)))
+                {
                     Directory.CreateDirectory(FileIO.StripFileName(filePath));
+                }
             }
             return filePath;
         }
 
-        /*public static string SelectDataFileSave(string iDataFile)
+        public static string GetConfigFilename()
         {
-            try
-            {
-                var str = Path.Combine(FPathAppData, iDataFile);
-                if (Debugger.IsAttached)
-                    str = SearchUp("Data", str);
-                if (!Directory.Exists(FileIO.StripFileName(str)))
-                    Directory.CreateDirectory(FileIO.StripFileName(str));
-                return str;
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show("Unable to create output folder: " + ex.Message);
-            }
+            if (File.Exists(FNameJsonConfig)) return FNameJsonConfig;
+            MessageBox.Show(@"Config file doesn't exist, generating a new one.", @"Missing config file", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            File.WriteAllText(FNameJsonConfig, "");
 
-            return string.Empty;
-        }*/
-
-        public static string GetConfigFilename(bool forceMhd)
-        {
-            switch (forceMhd)
-            {
-                case false when File.Exists(FNameJsonConfig):
-                {
-                    return FNameJsonConfig;
-                }
-                case false when !File.Exists(FNameJsonConfig):
-                {
-                    MessageBox.Show(@"Config file doesn't exist, generating a new one.", @"Missing config file", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    File.WriteAllText(FNameJsonConfig, "");
-                    return FNameJsonConfig;
-                }
-                case true when File.Exists(FNameConfig):
-                {
-                    return FNameConfig;
-                }
-            }
-
-            return FNameConfig;
-        }
-
-        internal static string SelectConfigFileSave()
-        {
-            try
-            {
-                if (!Directory.Exists(FileIO.StripFileName(FNameConfig)))
-                {
-                    Directory.CreateDirectory(FileIO.StripFileName(FNameConfig));
-                }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show("Unable to create output folder: " + ex.Message);
-            }
-
-            return FNameConfig;
+            return FNameJsonConfig;
         }
 
         public static class Headers
         {
             public const string VersionComment = "Version:";
 
-            internal static class DB
+            internal static class Db
             {
                 internal const string Start = "Mids Reborn Powers Database";
                 internal const string Archetypes = "BEGIN:ARCHETYPES";
@@ -216,7 +92,7 @@ namespace mrbBase
                 internal const string Summons = "BEGIN:SUMMONS";
             }
 
-            internal static class EnhDB
+            internal static class EnhDb
             {
                 internal const string Start = "Mids Reborn Enhancement Database";
             }
@@ -236,10 +112,17 @@ namespace mrbBase
                 internal const string Start = "Mids Reborn Attribute Modifier Tables";
             }
 
+            internal static class TypeGrade
+            {
+                internal const string Start = "Mids Reborn Types and Grades";
+            }
+
             public static class Save
             {
-                public const string Compressed = "MHDz";
-                public const string Uncompressed = "HeroDataVersion";
+                public const string Compressed = "MRBz";
+                public const string Uncompressed = "ToonDataVersion";
+                public const string LegacyCompressed = "MHDz";
+                public const string LegacyUncompressed = "HeroDataVersion";
             }
         }
